pyproject.toml
[tool.poetry]
name = "app"
version = "0.0.1"
description = "None"
authors = ["Danya <azathothdevk@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-telegram-bot = "^20.2"
SQLAlchemy = "^2.0.7"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
